feat: TransactionUserService 추상화 적용
