+++ app/src/main/java/com/techcourse/service/TxUserService.java
@@ -1,0 +1,39 @@
+package com.techcourse.service;
+
+import com.techcourse.domain.User;
+import nextstep.jdbc.DataAccessException;
+import org.springframework.transaction.PlatformTransactionManager;
+import org.springframework.transaction.support.DefaultTransactionDefinition;
+
+public class TxUserService implements UserService {
+
+    private final UserService userService;
+    private final PlatformTransactionManager transactionManager;
+
+    public TxUserService(final UserService userService, final PlatformTransactionManager transactionManager) {
+        this.userService = userService;
+        this.transactionManager = transactionManager;
+    }
+
+    @Override
+    public User findById(final long id) {
+        return userService.findById(id);
+    }
+
+    @Override
+    public void insert(final User user) {
+        userService.insert(user);
+    }
+
+    @Override
+    public void changePassword(final long id, final String newPassword, final String createBy) {
+        final var transactionStatus = transactionManager.getTransaction(new DefaultTransactionDefinition());
+        try {
+            userService.changePassword(id, newPassword, createBy);
+            transactionManager.commit(transactionStatus);
+        } catch (RuntimeException e) {
+            transactionManager.rollback(transactionStatus);
+            throw new DataAccessException(e);
+        }
+    }
+}
