+++ app/src/test/java/com/techcourse/service/UserServiceTest.java
@@ -12,6 +12,7 @@ import nextstep.jdbc.DataAccessException;
 import nextstep.jdbc.JdbcTemplate;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import org.springframework.jdbc.datasource.DataSourceTransactionManager;
 
 class UserServiceTest {
 
@@ -31,7 +32,9 @@ class UserServiceTest {
     @Test
     void testChangePassword() {
         final var userHistoryDao = new UserHistoryDao(jdbcTemplate);
-        final var userService = new UserService(userDao, userHistoryDao, jdbcTemplate.getDataSource());
+        final var appService = new AppUserService(userDao, userHistoryDao);
+        final var userService = new TxUserService(appService, new DataSourceTransactionManager(
+                jdbcTemplate.getDataSource()));
 
         final var newPassword = "qqqqq";
         final var createBy = "gugu";
@@ -46,7 +49,9 @@ class UserServiceTest {
     void testTransactionRollback() {
         // 트랜잭션 롤백 테스트를 위해 mock으로 교체
         final var userHistoryDao = new MockUserHistoryDao(jdbcTemplate);
-        final var userService = new UserService(userDao, userHistoryDao, jdbcTemplate.getDataSource());
+        final var appService = new AppUserService(userDao, userHistoryDao);
+        final var userService = new TxUserService(appService, new DataSourceTransactionManager(
+                jdbcTemplate.getDataSource()));
 
         final var newPassword = "newPassword";
         final var createBy = "gugu";
